#9509: removing an open .sla makes scribus crash
git-svn-id: svn://scribus.net/branches/Version135/Scribus@15945 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/scribus.cpp
@@ -361,7 +361,7 @@ int ScribusMainWindow::initScMW(bool primaryMainWindow)
 	if (scrActions["SaveAsDocumentTemplate"])
 		scrActions["SaveAsDocumentTemplate"]->setEnabled(false);
 
-	connect(ScCore->fileWatcher, SIGNAL(fileDeleted(QString )), this, SLOT(removeRecent(QString)));
+	connect(ScCore->fileWatcher, SIGNAL(fileDeleted(QString )), this, SLOT(removeRecentFromFileWatcher(QString)));
 	connect(this, SIGNAL(TextStyle(const ParagraphStyle&)), propertiesPalette, SLOT(updateStyle(const ParagraphStyle&)));
 	connect(this, SIGNAL(TextIFont(QString)), propertiesPalette, SLOT(setFontFace(QString)));
 	connect(this, SIGNAL(TextISize(double)), propertiesPalette, SLOT(setSize(double)));
@@ -3408,14 +3408,20 @@ void ScribusMainWindow::updateRecent(QString fn)
 	rebuildRecentFileMenu();
 }
 
-void ScribusMainWindow::removeRecent(QString fn)
+void ScribusMainWindow::removeRecent(QString fn, bool fromFileWatcher)
 {
 	if (RecentDocs.indexOf(fn) != -1)
 	{
 		RecentDocs.removeAll(fn);
-		ScCore->fileWatcher->removeFile(fn);
+		if (!fromFileWatcher)
+			ScCore->fileWatcher->removeFile(fn);
 	}
 	rebuildRecentFileMenu();
+}
+
+void ScribusMainWindow::removeRecentFromFileWatcher(QString fn)
+{
+	removeRecent(fn, true);
 }
 
 void ScribusMainWindow::loadRecent(QString fn)
